remove admin button
